Enable DWB to receive path updates (#409)
* Integrate example code into the DWB controller main loop
* Changes to get current robot pose from global planner, required for dynamic replanning
* Address review feedback
* Fix cpplint problem

+++ nav2_tasks/include/nav2_tasks/rate_controller_node.hpp
@@ -28,8 +28,8 @@ public:
   RateController(const std::string & name, const BT::NodeParameters & params)
   : BT::DecoratorNode(name, params)
   {
-    unsigned int hz = 1;
-    getParam<unsigned int>("hz", hz);
+    double hz = 1.0;
+    getParam<double>("hz", hz);
     period_ = 1.0 / hz;
   }
